Implemented Rule 184
Rule 184 is a one-dimension array of binary bits of data, where each bit
follows certain rules during each cycle. If a bit is 1, and its right
neighbor bit is 0, then the 1 bit moves to the right (if the neighbor is
1, then the 1 bit stays put). The current iteration uses 100 bit array,
and adds 75 1 bits randomly before running the cycles. The output of
each cycle is printed to a text to illustrate the process.